requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
colorlog==4.2.1
har2case==0.3.1
HttpRunner==1.5.8
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
jsonpath==0.82
lxml==4.2.5
MarkupSafe==1.1.1
more-itertools==8.4.0
pluggy==0.13.1
py==1.9.0
PyMySQL==0.9.3
pytest==4.5.0
pytest-html==1.19.0
pytest-metadata==1.10.0
PyYAML==3.12
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
